fit for each collection

+++ explanatory/plot_prob_rel.py
@@ -268,8 +268,8 @@ class PlotRelProb(object):
         if compact_x:
             xaxis = range(1, len(xaxis)+1)
 
-        xaxis = np.array(xaxis)
-        yaxis = np.array(yaxis)
+        xaxis = np.array(xaxis, dtype=np.float32)
+        yaxis = np.array(yaxis, dtype=np.float32)
         print xaxis, yaxis
         if curve_fitting and not plot_ratio:
             xaxis /= np.sum(xaxis)
